% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterStability.R
\name{ClusterStability}
\alias{ClusterStability}
\title{Compute cluster stability of MCAP output}
\usage{
ClusterStability(xx, k, B = 10, frac_subsample = 0.75)
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{k}{The number of clusters.}

\item{B}{The number of subsampled data sets to be used (default: \code{B=10}). 
Note that if \code{B} is too small (<10), the variance of the 
stability estimate increases.}

\item{frac_subsample}{Fraction of total samples to be used in each 
subsample (default: 0.75).}
}
\value{
\item{stab}{ Measure of cluster stability.}
}
\description{
Compute cluster stability w.r.t. pairs of data points over B random 
subsamples of size m < n (where n is the total sample size). Default for 
m=0.75n. Clustering is done using a full covariance Gaussian mixture model.
Measure of cluster stability is based on the adjusted Rand index
of shared points given two subsets of the data.
}
\examples{
  ## standard Normal matrix, 2 clusters:
  ClusterStability(xx = matrix(rnorm(1e3), 100), k = 2)
  
  \dontrun{
  ## large input matrix 
  ClusterStability(xx = matrix(rnorm(1e3*1e4), 1e3), k = 5, B = 20)
  }
}
\seealso{
\code{\link{MCAPfit}}, 
         \code{\link{OptDimClusterStability}}, 
         \code{\link{GMMwrapper}}, 
         \code{\link[mclust]{adjustedRandIndex}}
}
\author{
Bernd Taschler: \email{bernd.taschler@dzne.de}

Sach Mukherjee: \email{sach.mukherjee@dzne.de}
}
