% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMwrapper.R
\name{GMMwrapper}
\alias{GMMwrapper}
\title{Wrapper to do clustering with Gaussian mixture models}
\usage{
GMMwrapper(xx, k, criterion = "mmdl", true_labels = NULL,
  verbose = FALSE)
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{k}{The number of clusters.}

\item{criterion}{Optimisation criterion (`"bic"` or `"mmdl"`). Default: `"mmdl"`.}

\item{true_labels}{Vector of true cluster assignments (when provided, it is 
used to compute the Rand index).}

\item{verbose}{Logical, when true: print progress information.}
}
\value{
@param model_fit Model fit (output of `mixglasso()`).

@param aRI Adjusted Rand index (when `true_labels` are provided).
}
\description{
Full covariance Gaussian mixture modelling based on the nethet package. 
The wrapper performs multiple restarts in case true labels are provided 
(in order to optimise cluster assignments w.r.t. the Rand index).
}
\seealso{
\code{\link{runMCAP}}

\code{\link{OptDimClusterStability}}
}
\author{
Bernd Taschler \email{bernd.taschler@dzne.de}

Sach Mukherjee \email{sach.mukherjee@dzne.de}
}
