% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GramPCA.R
\name{GramPCA}
\alias{GramPCA}
\title{Perform PCA using Gram matrix}
\usage{
GramPCA(xx, npc)
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{npc}{The number of principal components to be returned.}
}
\value{
\item{zz}{Principal components (n x npc).}
}
\description{
This function performs PCA using the eigenvectors of the Gram matrix G 
(G=X*X') when n < p, or standard PCA when n > p.
}
\examples{
  ## standard Normal 10x50 matrix, 3 PC's
  GramPCA(xx = matrix(rnorm(500),10), npc = 3)
  
  ## 100x5 matrix (use standard PCA):
  GramPCA(xx = matrix(rnorm(500),100), npc = 3)
  
  \dontrun{
  ## large 100x10,000 matrix:
  GramPCA(xx = matrix(rnorm(1e6),100), npc = 3)
  }
}
\seealso{
\code{\link[pcaMethods]{pca}}

\code{\link{eigen}}
}
\author{
Bernd Taschler: \email{bernd.taschler@dzne.de}
}
