% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMwrapper.R
\name{KMwrapper}
\alias{KMwrapper}
\title{Wrapper to do K-means clustering}
\usage{
KMwrapper(xx, k, true_labels = NULL, verbose = FALSE)
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{k}{The number of clusters.}

\item{true_labels}{Vector of true cluster assignments (when provided, it is 
used to compute the Rand index).}

\item{verbose}{Logical, when true: print progress information.}
}
\value{
\item{model_fit}{ Model fit (output of \code{\link[stats]{kmeans}}).}

\item{aRI}{ Adjusted Rand index (when \code{true_labels} is provided).}
}
\description{
Standard K-means clustering. 
The wrapper performs multiple restarts in case true labels are provided 
(in order to optimise cluster assignments w.r.t. the Rand index).
}
\seealso{
\code{\link[stats]{kmeans}}, \code{\link[mclust]{adjustedRandIndex}}
}
