% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCAPfit.R
\name{MCAPfit}
\alias{MCAPfit}
\title{Model based clustering via adaptive (linear) projections}
\usage{
MCAPfit(xx, k, projection = "PCA", true_labels = NULL,
  centering_per_group = FALSE, parallel = FALSE, verbose = FALSE,
  ...)
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{k}{The number of clusters.}

\item{projection}{Projection method (\code{"PCA"}, \code{"gaussian"}, 
\code{"achlioptas"} or \code{"li"}). Default: \code{"PCA"}.}

\item{true_labels}{Vector of true cluster assignments (when provided, it is 
used to compute the Rand index).}

\item{centering_per_group}{Logical, when true: mean centre input matrix (if true 
labels are provided: centre data per group)}

\item{parallel}{Logical, when true: perform line search over projection 
dimension in parallel.}

\item{verbose}{Logical, when true: print some progress information.}

\item{...}{Additional options for \code{\link{OptDimClusterStability}} and
\code{\link{GMMwrapper}}.}
}
\value{
\item{fit_gmm}{ Model fit (GMM output of \code{\link[nethet]{mixglasso}}), 
                        including BIC, MMDL and adj. Rand index (when 
                        \code{true_labels} is provided).}
        \item{fit_q_opt}{ Output of \code{\link{OptDimClusterStability}}.}
}
\description{
\code{"MCAPfit"} performs model based clustering using full variance 
Gaussian mixtures in a lower dimensional projected space obtained via 
adaptive (linear) projections.
Projection variants include PCA-based and random projection.
}
\details{
...
}
\examples{
  ## small standard Normal matrix with random labels:
  MCAPfit(xx=matrix(rnorm(200),20,10), k=2, projection='PCA', 
          true_labels = round(runif(20)), parallel=TRUE)
  
  \dontrun{
  ## sparse random projection:
  MCAPfit(xx=rbind(matrix(rnorm(5e4),100,500), matrix(rnorm(5e4,mean=1),100,500)),
          k=2, projection='li, 
          true_labels=c(rep(0,100),rep(1,100)), parallel=TRUE)
          
  ## sparse random projection, removing the mean signal:
  MCAPfit(xx=rbind(matrix(rnorm(5e4),100,500), matrix(rnorm(5e4,mean=1),100,500)),
          k=2, projection='li, centering_per_group=TRUE,
          true_labels=c(rep(0,100),rep(1,100)), parallel=TRUE)
  } 
}
\references{
Taschler, B., Dondelinger, F. and Mukherjee, S. (2019) 
            Model based clustering via adaptive projections \url{https://arxiv.org/pdf/??.pdf}
}
\seealso{
\code{\link{GMMwrapper}}, 
         \code{\link{OptDimClusterStability}},
         \code{\link{ClusterStability}}
}
\author{
Bernd Taschler: \email{bernd.taschler@dzne.de}

Sach Mukherjee: \email{sach.mukherjee@dzne.de}
}
