% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precision2PartialCor.R
\name{Precision2PartialCor}
\alias{Precision2PartialCor}
\title{Compute partial correlation matrix from inverse covariance matrix}
\usage{
Precision2PartialCor(xx)
}
\arguments{
\item{xx}{The inverse covariance matrix.}
}
\value{
\item{pcor}{ The partial correlation matrix}.
}
\description{
This function takes an inverse covariance (or precision) matrix as input 
and computes the partial correlation matrix.
}
