% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCAP.R
\name{runMCAP}
\alias{runMCAP}
\title{Model based clustering via adaptive (linear) projections}
\usage{
runMCAP(xx, k, projection = "PCA", true_labels = NULL,
  centering_per_group = FALSE, parallel = FALSE, verbose = FALSE,
  ...)
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{k}{The number of clusters.}

\item{projection}{Projection method (`"PCA"`, `"gaussian"`, `"achlioptas"` 
or `"li"`). Default: `"PCA"`.}

\item{true_labels}{Vector of true cluster assignments (when provided, it is 
used to compute the Rand index).}

\item{centering_per_group}{Logical, when true: mean centre input matrix (if true 
labels are provided: centre data per group)}

\item{parallel}{Logical, when true: perform line search over projection 
dimension in parallel.}

\item{verbose}{Logical, when true: print some progress information.}

\item{...}{Additional options for `OptDimClusterStability()` and `GMMwrapper()`.}
}
\value{
@param fit_gmm Model fit (GMM output of `mixglasso()`), including
               BIC, MMDL and adj. Rand index (when `true_labels` are provided).

@param fit_q_opt Output of `OptDimClusterStability()`.
}
\description{
Model based clustering using full variance Gaussian mixtures in a lower
dimensional projected space obtained via adaptive (linear) projections.
Projection variants include PCA-based and random projection.
}
\references{
Taschler, B., Dondelinger, F. and Mukherjee, S. (2019) 
            Model based clustering via adaptive projections \url{https://arxiv.org/pdf/??.pdf}
}
\seealso{
\code{\link{GMMwrapper}}

\code{\link{OptDimClusterStability}}
}
\author{
Bernd Taschler \email{bernd.taschler@dzne.de}

Sach Mukherjee \email{sach.mukherjee@dzne.de}
}
