% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECTRALwrapper.R
\name{SPECTRALwrapper}
\alias{SPECTRALwrapper}
\title{Wrapper to do spectral clustering}
\usage{
SPECTRALwrapper(xx, k, true_labels = NULL, verbose = FALSE)
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{k}{The number of clusters.}

\item{true_labels}{Vector of true cluster assignments (when provided, it is 
used to compute the Rand index).}

\item{verbose}{Logical, when true: print progress information.}
}
\value{
@param model_fit Model fit (output of `mixglasso()`).

@param aRI Adjusted Rand index (when `true_labels` are provided).
}
\description{
Spectral clustering using an RBF kernel. 
The wrapper performs multiple restarts in case true labels are provided 
(in order to optimise cluster assignments w.r.t. the Rand index).
}
