% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMwrapper.R
\name{GMMwrapper}
\alias{GMMwrapper}
\title{Wrapper to do clustering with Gaussian mixture models}
\usage{
GMMwrapper(xx, k, criterion = "mmdl", true_labels = NULL,
  verbose = FALSE)
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{k}{The number of clusters.}

\item{criterion}{Optimisation criterion (\code{"bic"} or \code{"mmdl"}). 
Default: \code{"mmdl"}.}

\item{true_labels}{Vector of true cluster assignments (when provided, it is 
used to compute the Rand index).}

\item{verbose}{Logical, when true: print progress information.}
}
\value{
\item{model_fit}{ Model fit (output of \code{\link[nethet]{mixglasso}}).}
        \item{aRI}{ Adjusted Rand index (when \code{true_labels} is provided).}
}
\description{
Full covariance Gaussian mixture modelling based on the nethet package. 
The wrapper performs multiple restarts in case true labels are provided 
(in order to optimise cluster assignments w.r.t. the Rand index).
}
\examples{
  ## standard Normal 100x5 matrix, 2 clusters:
  GMMwrapper(xx = matrix(rnorm(500),100), k = 2) 
}
\seealso{
\code{\link{MCAPfit}},
         \code{\link{OptDimClusterStability}},
         \code{\link{ClusterStability}}

\code{\link[mclust]{adjustedRandIndex}}
}
\author{
Bernd Taschler: \email{bernd.taschler@dzne.de}
}
