% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GramPCA.R
\name{GramPCA}
\alias{GramPCA}
\title{Perform PCA using Gram matrix}
\usage{
GramPCA(xx, npc)
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{npc}{The number of principal components to be returned.}
}
\value{
@param zz Principal components (n x npc).
}
\description{
This function performs PCA using the eigenvectors of the Gram matrix G 
(G=X*X') when n < p, or standard PCA when n > p.
}
