% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCLUSTwrapper.R
\name{HCLUSTwrapper}
\alias{HCLUSTwrapper}
\title{Wrapper to do hierarchical clustering}
\usage{
HCLUSTwrapper(xx, k, method = "ward.D", true_labels = NULL,
  verbose = FALSE)
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{k}{The number of clusters.}

\item{method}{Clustering method (see \code{\link[stats]{hclust}}). 
If \code{method = "all"} and \code{true_labels} is provided: 
optimise clustering w.r.t. Rand index  and return method 
with highest adjusted RI. Default: \code{method = "Ward.D"}.}

\item{true_labels}{Vector of true cluster assignments (when provided, it is 
used to compute the Rand index).}

\item{verbose}{Logical, when true: print progress information.}
}
\value{
\item{model_fit}{ Model fit (output of \code{\link[stats]{hclust}}).}

\item{aRI}{ Adjusted Rand index (when \code{true_labels} is provided).}
}
\description{
Hierarchical clustering using Euclidean distances and (by default) 
the Ward criterion. 
The wrapper can perform an optimisation over clustering methods in case 
true labels are provided (i.e. optimising cluster assignments w.r.t. the Rand index).
}
\seealso{
\code{\link[stats]{hclust}}, \code{\link[mclust]{adjustedRandIndex}}
}
\author{
Bernd Taschler: \email{bernd.taschler@dzne.de}
}
