% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptDimClusterStability.R
\name{OptDimClusterStability}
\alias{OptDimClusterStability}
\title{Determine optimal projection dimension (PCA or random projections) based
on cluster stability}
\usage{
OptDimClusterStability(xx, k, method = "PCA", n_grid = 5,
  q_max = min(ncol(xx), sqrt(10 * nrow(xx)/k)), true_labels = NULL,
  parallel = FALSE, verbose = FALSE)
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{k}{The number of clusters.}

\item{method}{Projection method (\code{"PCA"} or random projections: 
\code{"gaussian"}, \code{"achlioptas"} or \code{"li"}). 
Default: \code{"PCA"}.}

\item{n_grid}{Number of values to be used in the line search for optimal
projection dimension. Default: 5.}

\item{q_max}{Maximum target dimension to be used in line search. (Note: 
the smallest target dimension is always \code{k}, 
the maximum may not exceed the total dimensionality p). 
Default: \code{sqrt(10n/k)}.}

\item{true_labels}{Vector of true cluster assignments (if provided, it is 
used to compute the Rand index and \code{q_star}).}

\item{parallel}{Logical, if true: perform line search over \code{q} in parallel.}

\item{verbose}{Logical, if true: print progress information.}
}
\value{
\item{q_opt}{ Optimal target dimension (maximises cluster stability).}
        \item{stab_score}{ Stability measure for \code{q_opt}.}
        \item{q_star}{ Optimal ("oracle") target dimension (maximises adj. 
                       Rand index). Only available if true labels have been 
                       provided.}
}
\description{
Find the optimal projection dimension for PCA or random projections 
based on cluster stability by performing a line search over the target 
dimension q.
}
\examples{
  ## Default settings, 50x10 standard Normal input matrix:
  OptDimClusterStability(xx=matrix(rnorm(500),50), k=2)
  
  ## finer search over q:
  OptDimClusterStability(xx=matrix(rnorm(2e4),100,200), k=2, n_grid=10)
  
  ## set max. q, provide class labels, run in parallel:
  \dontrun{
  OptDimClusterStability(xx=rbind(matrix(rnorm(2e4),100,200), 
                                  matrix(rnorm(2e4, mean = 2),100,200)),
                         k=2, q_max=15, true_labels = c(rep(0,100), rep(1,100)),
                         parallel=TRUE)
 }
}
\seealso{
\code{\link{MCAPfit}},
         \code{\link{GMMwrapper}}, 
         \code{\link{ClusterStability}}
}
\author{
Bernd Taschler \email{bernd.taschler@dzne.de}

Sach Mukherjee \email{sach.mukherjee@dzne.de}
}
