% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandProject.R
\name{RandProject}
\alias{RandProject}
\title{Random projection of input matrix X}
\usage{
RandProject(xx, q, method = "gaussian")
}
\arguments{
\item{xx}{The data matrix (n x p).}

\item{q}{The target dimension.}

\item{method}{Projection method using either a Gaussian (\code{"gaussian"}), 
a sparse (\code{"achlioptas"}) or very sparse (\code{"li"}) projection 
matrix. Default: \code{"gaussian"}.}
}
\value{
Projected matrix (n x q).
}
\description{
Perform a random projection (Gauss, Achlioptas or Li) of the input matrix X.
Based on the RandPro package.
}
\seealso{
\code{\link[RandPro]{form_matrix}}
}
