% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rank2Normality.R
\name{Rank2Normality}
\alias{Rank2Normality}
\title{Transform data matrix to normality based on rank}
\usage{
Rank2Normality(xx)
}
\arguments{
\item{xx}{The data matrix.}
}
\value{
tt Data matrix transformed to normality.
}
\description{
This function performs a rank to normality transformation of the columns
of an input matrix X.
}
\seealso{
\code{\link[stats]{qnorm}}
}
